fastindex: hopefully fix bug with one of the other stream files appearing late. Changes didn't seem to brake anything, but it is hard to test this live mode behavior
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@9534 b967ad99-d558-0410-b138-e0f6c56caec7

+++ Translator/branches/fastindex/include/H5MpiTranslateApp.h
@@ -192,7 +192,7 @@ private:
   int64_t m_numEventsToCheckForDoneWorkers;
 
   bool m_fastIndex;
-  int m_fastIndexMBhalfBlock;
+  double m_fastIndexMBhalfBlock;
   int m_fastIndexNumberBlocksToTry;
 
   LusiTime::Time m_startTime, m_endTime;
